requirements.txt
Flask==2.0.2
python-dotenv
requests
pycryptodome
Werkzeug<3
pandas
openpyxl